refactor: move utils from components directory
