+++ src/utils/eventListenerManager.ts
