+++ src/utils/locationShortener.ts
@@ -1,0 +1,4 @@
+export const getClassroomInfo = (location: string): string => {
+    const classroom = location.split(" ")[0];
+    return classroom;
+};
